move refund logic to gas calculator, delete RefundCalculator
