+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/mainnet/ClassicProtocolSpecs.java
@@ -171,14 +171,12 @@ public class ClassicProtocolSpecs {
                     evm, true, Collections.singletonList(MaxCodeSizeRule.from(evm)), 1))
         .transactionProcessorBuilder(
             (gasCalculator,
-                refundCalculator,
                 feeMarket,
                 transactionValidatorFactory,
                 contractCreationProcessor,
                 messageCallProcessor) ->
                 new MainnetTransactionProcessor(
                     gasCalculator,
-                    refundCalculator,
                     transactionValidatorFactory,
                     contractCreationProcessor,
                     messageCallProcessor,
@@ -355,14 +353,12 @@ public class ClassicProtocolSpecs {
         // EIP-3651
         .transactionProcessorBuilder(
             (gasCalculator,
-                refundCalculator,
                 feeMarket,
                 transactionValidatorFactory,
                 contractCreationProcessor,
                 messageCallProcessor) ->
                 new MainnetTransactionProcessor(
                     gasCalculator,
-                    refundCalculator,
                     transactionValidatorFactory,
                     contractCreationProcessor,
                     messageCallProcessor,
