--- evm/src/main/java/org/hyperledger/besu/evm/refundcalculator/PragueRefundCalculator.java
@@ -1,38 +1,0 @@
-/*
- * Copyright contributors to Hyperledger Besu.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-package org.hyperledger.besu.evm.refundcalculator;
-
-import org.hyperledger.besu.datatypes.Transaction;
-import org.hyperledger.besu.evm.frame.MessageFrame;
-import org.hyperledger.besu.evm.gascalculator.GasCalculator;
-
-/** Refund calculator for the Prague hard fork. */
-public class PragueRefundCalculator extends FrontierRefundCalculator implements RefundCalculator {
-
-  /** Instantiates a new Prague refund calculator. */
-  public PragueRefundCalculator() {
-    // empty
-  }
-
-  @Override
-  // TODO: Implement the calculateGasRefund method for Prague, this is a simple placeholder
-  public long calculateGasRefund(
-      final GasCalculator gasCalculator,
-      final Transaction transaction,
-      final MessageFrame initialFrame,
-      final long codeDelegationRefund) {
-    return super.calculateGasRefund(gasCalculator, transaction, initialFrame, codeDelegationRefund);
-  }
-}
